services/document-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1
groq==0.11.0
PyJWT==2.8.0
pypdf==4.2.0
jsonschema==4.22.0
structlog==24.1.0
python-dotenv==1.0.1
httpx==0.27.0

# Test dependencies
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
pytest-benchmark==4.0.0
pytest-mock==3.14.0
Faker==25.2.0
factory-boy==3.3.0
